+++ Cesium3DTilesSelection/src/Tile.cpp
@@ -540,6 +540,11 @@ void Tile::processLoadedContent() {
       if (!this->_pContent->model) {
         this->setUnconditionallyRefine();
       }
+      // else if it has a model, try to get the copyright
+      else if (this->_pContent->model->asset.copyright) {
+        this->_pContent->credit = externals.pCreditSystem->createCredit(
+            *this->_pContent->model->asset.copyright);
+      }
 
       // A new and improved bounding volume.
       if (this->_pContent->updatedBoundingVolume) {