+++ Cesium3DTilesSelection/src/Tileset.cpp
@@ -315,6 +315,11 @@ Tileset::updateView(const std::vector<ViewState>& frustums) {
           }
         }
       }
+
+      if (tile->getContent() != nullptr &&
+          tile->getContent()->credit.has_value()) {
+        pCreditSystem->addCreditToFrame(*tile->getContent()->credit);
+      }
     }
   }