+++ CHANGES.md
@@ -10,6 +10,7 @@
 
 - Added in-memory cache for Ion asset endpoint responses to avoid repeated requests.
 - Add ScopeGuard utility to automatically execute function when exiting a scope.
+- Add glTF copyright information to the list of tileset credits.
 
 ### v0.13.0 - 2022-03-01